# public/openapi.yaml
openapi: 3.1.0
info:
  title: YouTube Summaries API
  description: |
    API documentation for YouTube Summaries - AI-powered video transcript extraction and summarization.

    ## API Versioning

    All public API endpoints are now versioned under `/api/v1/*` for future-proofing:
    - **Versioned endpoints** (recommended): `/api/v1/summarize`, `/api/v1/chat`, etc.
    - **Unversioned endpoints** (deprecated): `/api/summarize`, `/api/chat`, etc. (proxy to v1)

    This versioning strategy allows backwards compatibility while enabling API evolution.

    ## API Layers

    This documentation covers three API layers:
    1. **Next.js API Routes** - Main application endpoints (`/api/v1/*`)
    2. **Cloudflare Worker** - D1 caching layer for transcripts
    3. **fast-proxy-api** - Source for transcript/metadata (see API_REFERENCE.md)

    ## Authentication

    - Most endpoints are public with IP-based rate limiting
    - Authenticated endpoints require Clerk session cookies
    - Webhook endpoints use signature verification

    ## Rate Limiting

    All endpoints return rate limit headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets
  version: 2.0.0
  contact:
    name: YouTube Summaries
    url: https://youtubesummaries.cc
  license:
    name: MIT
servers:
  - url: https://youtubesummaries.cc
    description: Production
...
    description: Development
...
paths:
...
  /api/v1/youtube/search:
    get:
      tags: [YouTube]
      summary: Search YouTube videos
      description: |
        Search for YouTube videos with multi-tier fallback:
        1. fast-proxy-api backend (preferred)
        2. Direct YouTube Data API v3 (if API key configured)
        3. Demo videos (final fallback)

        **Rate Limit:** 30 requests/minute
      operationId: searchVideosV1
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
          example: "machine learning tutorial"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              examples:
                success:
                  summary: Live results
                  value:
                    videos:
                      - id: "dQw4w9WgXcQ"
                        title: "Rick Astley - Never Gonna Give You Up"
                        thumbnail: "https://i.ytimg.com/vi/dQw4w9WgXcQ/mqdefault.jpg"
                        channelTitle: "Rick Astley"
                        publishedAt: "2009-10-25T06:57:33Z"
                demo:
                  summary: Demo fallback
                  value:
                    videos: []
                    isDemo: true
                    warning: "YouTube API quota exceeded. Showing demo videos."
        '400':
          description: Missing query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Query parameter is required"
...
  /api/v1/beta:
    post:
      tags: [Auth]
      summary: Join beta waitlist
      description: |
        Register email for beta access waitlist.

        **Rate Limit:** 5 requests/hour per IP
      operationId: joinBetaV1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BetaSignupRequest'
            example:
              email: "user@example.com"
              source: "homepage"
      responses:
        '201':
          description: Successfully joined waitlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BetaSignupResponse'
              example:
                message: "Successfully joined the beta waitlist!"
        '200':
          description: Already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BetaSignupResponse'
              example:
                message: "You're already on the list! We'll be in touch soon."
        '400':
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Please enter a valid email address"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'

  # ===========================================
  # LEGACY UNVERSIONED ENDPOINTS (Deprecated)
  # ===========================================
  # These endpoints proxy to /api/v1/* for backwards compatibility
  # Clients should migrate to v1 endpoints
...
  /api/youtube/search:
    get:
      tags: [YouTube]
      summary: Search YouTube videos (deprecated)
      deprecated: true
      description: |
        **DEPRECATED:** Use `/api/v1/youtube/search` instead. This endpoint proxies to v1 for backwards compatibility.

        Search for YouTube videos with multi-tier fallback:
        1. fast-proxy-api backend (preferred)
        2. Direct YouTube Data API v3 (if API key configured)
        3. Demo videos (final fallback)

        **Rate Limit:** 30 requests/minute
      operationId: searchVideos
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
          example: "machine learning tutorial"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              examples:
                success:
                  summary: Live results
                  value:
                    videos:
                      - id: "dQw4w9WgXcQ"
                        title: "Rick Astley - Never Gonna Give You Up"
                        thumbnail: "https://i.ytimg.com/vi/dQw4w9WgXcQ/mqdefault.jpg"
                        channelTitle: "Rick Astley"
                        publishedAt: "2009-10-25T06:57:33Z"
                demo:
                  summary: Demo fallback
                  value:
                    videos: []
                    isDemo: true
                    warning: "YouTube API quota exceeded. Showing demo videos."
        '400':
          description: Missing query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Query parameter is required"
...
  /api/beta:
    post:
      tags: [Auth]
      summary: Join beta waitlist (deprecated)
      deprecated: true
      description: |
        **DEPRECATED:** Use `/api/v1/beta` instead. This endpoint proxies to v1 for backwards compatibility.

        Register email for beta access waitlist.

        **Rate Limit:** 5 requests/hour per IP
      operationId: joinBeta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BetaSignupRequest'
            example:
              email: "user@example.com"
              source: "homepage"
      responses:
        '201':
          description: Successfully joined waitlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BetaSignupResponse'
              example:
                message: "Successfully joined the beta waitlist!"
        '200':
          description: Already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BetaSignupResponse'
              example:
                message: "You're already on the list! We'll be in touch soon."
        '400':
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Please enter a valid email address"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
components:
  securitySchemes:
    clerkAuth:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk session cookie
    polarWebhookSignature:
      type: apiKey
      in: header
      name: X-Polar-Signature
      description: Polar.sh webhook signature
  headers:
    X-RateLimit-Limit:
      description: Maximum requests allowed in window
      schema:
        type: integer
    X-RateLimit-Remaining:
      description: Requests remaining in current window
      schema:
        type: integer
    X-RateLimit-Reset:
      description: Unix timestamp when rate limit resets
      schema:
        type: integer
  schemas:
...
    BetaSignupRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: Email address
        source:
          type: string
          default: website
          description: Signup source for analytics
...
    SearchResponse:
      type: object
      properties:
        videos:
          type: array
          items:
            $ref: '#/components/schemas/VideoSummary'
        isDemo:
          type: boolean
          description: True if returning demo data due to API issues
        warning:
          type: string
          description: Warning message when using fallback data
    VideoSummary:
      type: object
      properties:
        id:
          type: string
          description: YouTube video ID
        title:
          type: string
        thumbnail:
          type: string
          format: uri
        channelTitle:
          type: string
        publishedAt:
          type: string
          format: date-time
...
    BetaSignupResponse:
      type: object
      properties:
        message:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
...
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          example: "Rate limit exceeded. Please try again later."
        retryAfter:
          type: integer
          description: Seconds until rate limit resets
  responses:
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'
...